LLM/LangChain/semantic_search/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
# created-by: conda 25.5.1
aiohappyeyeballs=2.6.1=pyhd8ed1ab_0
aiohttp=3.12.14=py312h6daa0e5_0
aiosignal=1.4.0=pyhd8ed1ab_0
annotated-types=0.7.0=pyhd8ed1ab_1
anyio=4.9.0=pyh29332c3_0
async-timeout=4.0.3=pyhd8ed1ab_0
attrs=25.3.0=pyh71513ae_0
brotli-python=1.1.0=py312hd8f9ff3_3
bzip2=1.0.8=h99b78c6_7
ca-certificates=2025.7.9=hbd8a1cb_0
certifi=2025.7.9=pyhd8ed1ab_0
cffi=1.17.1=py312h0fad829_0
charset-normalizer=3.4.2=pyhd8ed1ab_0
dataclasses-json=0.6.7=pyhd8ed1ab_1
exceptiongroup=1.3.0=pyhd8ed1ab_0
frozenlist=1.7.0=py312h512c567_0
greenlet=3.2.3=py312hd8f9ff3_0
h11=0.16.0=pyhd8ed1ab_0
h2=4.2.0=pyhd8ed1ab_0
hpack=4.1.0=pyhd8ed1ab_0
httpcore=1.0.9=pyh29332c3_0
httpx=0.28.1=pyhd8ed1ab_0
httpx-sse=0.4.1=pyhd8ed1ab_0
hyperframe=6.1.0=pyhd8ed1ab_0
idna=3.10=pyhd8ed1ab_1
jsonpatch=1.33=pyhd8ed1ab_1
jsonpointer=3.0.0=py312h81bd7bf_1
langchain=0.3.26=pymin312_hff2d567_0
langchain-community=0.3.27=pymin312_hff2d567_0
langchain-core=0.3.68=pyhd8ed1ab_0
langchain-text-splitters=0.3.8=pyhd8ed1ab_0
langsmith=0.3.45=pyhd8ed1ab_0
libblas=3.9.0=32_h10e41b3_openblas
libcblas=3.9.0=32_hb3479ef_openblas
libcxx=20.1.8=ha82da77_0
libexpat=2.7.0=h286801f_0
libffi=3.4.6=h1da3d7d_1
libgfortran=5.0.0=14_2_0_h6c33f7e_103
libgfortran5=14.2.0=h6c33f7e_103
liblapack=3.9.0=32_hc9a63f6_openblas
liblzma=5.8.1=h39f12f2_2
libopenblas=0.3.30=openmp_hf332438_0
libsqlite=3.50.2=hf8de324_2
libzlib=1.3.1=h8359307_2
llvm-openmp=20.1.7=hdb05f8b_0
marshmallow=3.26.1=pyhd8ed1ab_0
multidict=6.6.3=py312hdb8e49c_0
mypy_extensions=1.1.0=pyha770c72_0
ncurses=6.5=h5e97a16_3
numpy=1.26.4=py312h8442bc7_0
openssl=3.5.1=h81ee809_0
orjson=3.10.18=py312hd60eec9_1
packaging=24.2=pyhd8ed1ab_2
pip=25.1.1=pyh8b19718_0
propcache=0.3.1=py312h998013c_0
pycparser=2.22=pyh29332c3_1
pydantic=2.11.7=pyh3cfb1c2_0
pydantic-core=2.33.2=py312hd3c0895_0
pydantic-settings=2.10.1=pyh3cfb1c2_0
pypdf=5.7.0=pyhe01879c_0
pysocks=1.7.1=pyha55dd90_7
python=3.12.11=hc22306f_0_cpython
python-dotenv=1.1.1=pyhe01879c_0
python_abi=3.12=7_cp312
pyyaml=6.0.2=py312h998013c_2
readline=8.2=h1d1bf99_2
requests=2.32.4=pyhd8ed1ab_0
requests-toolbelt=1.0.0=pyhd8ed1ab_1
setuptools=80.9.0=pyhff2d567_0
sniffio=1.3.1=pyhd8ed1ab_1
sqlalchemy=2.0.41=py312hea69d52_0
tenacity=9.1.2=pyhd8ed1ab_0
tk=8.6.13=h892fb3f_2
typing-extensions=4.14.1=h4440ef1_0
typing-inspection=0.4.1=pyhd8ed1ab_0
typing_extensions=4.14.1=pyhe01879c_0
typing_inspect=0.9.0=pyhd8ed1ab_1
tzdata=2025b=h78e105d_0
urllib3=2.5.0=pyhd8ed1ab_0
wheel=0.45.1=pyhd8ed1ab_1
yaml=0.2.5=h3422bc3_2
yarl=1.20.1=py312h998013c_0
zstandard=0.23.0=py312hea69d52_2
